requirements.txt
beautifulsoup4==4.12.2
GitPython==3.1.31
langchain~=0.0.137
mypy==1.2.0
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.4
pandas==2.0.0
pandas-stubs==1.5.3.230321
PyGithub==1.58.1
pytest==7.2.2
python-dotenv==1.0.0
regex~=2022.10.31
requests==2.28.2
types-beautifulsoup4==4.12.0.3
types-regex==2023.3.23.1
